// frontend/src/app/core/api/model/userResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponse { 
    id?: number;
    email?: string;
    username?: string;
    name?: string;
    phoneNumber?: string;
    createdAt?: string;
    is2faEnabled?: boolean;
}